+++ api/face_classify.py
@@ -26,7 +26,7 @@ def cluster_faces(user):
     faces = Face.objects.filter(photo__owner=user)
     face_encodings_all = []
     for face in faces:
-        face_encoding = np.frombuffer(bytes.fromhex(face.encoding))
+        face_encoding = np.frombuffer(bytes.fromhex(face.encoding), dtype=np.float32 )
         face_encodings_all.append(face_encoding)
 
     pca = PCA(n_components=3)
@@ -70,7 +70,7 @@ def train_faces(user, job_id):
         persons_list = []
         person_face_encodings = []
         for person in persons:
-            face_encoding = np.frombuffer(bytes.fromhex(person.mean_face_encoding))
+            face_encoding = np.frombuffer(bytes.fromhex(person.mean_face_encoding), dtype=np.float32 )
             person_face_encodings.append(face_encoding)
             persons_list.append(person)
         faces = Face.objects.filter(
@@ -87,7 +87,7 @@ def train_faces(user, job_id):
         target_count = len(faces)
         for idx, face in enumerate(faces):
             if face.person_label_is_inferred in [True, None]:
-                face_encoding = np.frombuffer(bytes.fromhex(face.encoding))
+                face_encoding = np.frombuffer(bytes.fromhex(face.encoding), dtype=np.float32)
                 results = face_recognition.face_distance(person_face_encodings, face_encoding)
                 TOLERANCE = 0.5
                 person_idx = np.argmin(results)
